refactor mock jwt and userValidation
