+++ routes/jwt.js
@@ -1,0 +1,7 @@
+//Mock jwt auth
+//Delete when jwt is ready
+const verifyJsonWebToken = (token, callback) => {
+  return callback(null, { password: '1234567' });
+};
+
+module.exports = verifyJsonWebToken;
